requirements.txt
# Core FastAPI dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.20

# Deep Learning frameworks
torch>=2.5.0
torchvision>=0.20.0
ultralytics>=8.3.0

# Image processing
opencv-python-headless>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# Data validation
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=24.0.0

# Training dependencies
tqdm>=4.66.3
matplotlib>=3.8.2
seaborn>=0.13.0
pyyaml>=6.0
